+++ llama_index/llms/base.py
@@ -8,6 +8,8 @@ from llama_index.constants import DEFAULT_CONTEXT_WINDOW, DEFAULT_NUM_OUTPUTS
 
 
 class MessageRole(str, Enum):
+    """Message role."""
+
     SYSTEM = "system"
     USER = "user"
     ASSISTANT = "assistant"
@@ -16,6 +18,8 @@ class MessageRole(str, Enum):
 
 # ===== Generic Model Input - Chat =====
 class ChatMessage(BaseModel):
+    """Chat message."""
+
     role: MessageRole = MessageRole.USER
     content: Optional[str] = ""
     additional_kwargs: dict = Field(default_factory=dict)
@@ -26,6 +30,8 @@ class ChatMessage(BaseModel):
 
 # ===== Generic Model Output - Chat =====
 class ChatResponse(BaseModel):
+    """Chat response."""
+
     message: ChatMessage
     raw: Optional[dict] = None
     delta: Optional[str] = None
@@ -38,6 +44,8 @@ ChatResponseGen = Generator[ChatResponse, None, None]
 
 # ===== Generic Model Output - Completion =====
 class CompletionResponse(BaseModel):
+    """Completion response."""
+
     text: str
     additional_kwargs: dict = Field(default_factory=dict)
     raw: Optional[dict] = None
@@ -55,30 +63,38 @@ class LLMMetadata(BaseModel):
 
     context_window: int = DEFAULT_CONTEXT_WINDOW
     num_output: int = DEFAULT_NUM_OUTPUTS
+    is_chat_model: bool = False
 
 
 class LLM(ABC):
+    """LLM interface."""
+
     @property
     @abstractmethod
     def metadata(self) -> LLMMetadata:
+        """LLM metadata."""
         pass
 
     @abstractmethod
     def chat(self, messages: Sequence[ChatMessage], **kwargs: Any) -> ChatResponse:
+        """Chat endpoint for LLM."""
         pass
 
     @abstractmethod
     def complete(self, prompt: str, **kwargs: Any) -> CompletionResponse:
+        """Completion endpoint for LLM."""
         pass
 
     @abstractmethod
     def stream_chat(
         self, messages: Sequence[ChatMessage], **kwargs: Any
     ) -> ChatResponseGen:
+        """Streaming chat endpoint for LLM."""
         pass
 
     @abstractmethod
     def stream_complete(self, prompt: str, **kwargs: Any) -> CompletionResponseGen:
+        """Streaming completion endpoint for LLM."""
         pass
 
     # ===== Async Endpoints =====
@@ -86,20 +102,24 @@ class LLM(ABC):
     async def achat(
         self, messages: Sequence[ChatMessage], **kwargs: Any
     ) -> ChatResponse:
+        """Async chat endpoint for LLM."""
         pass
 
     @abstractmethod
     async def acomplete(self, prompt: str, **kwargs: Any) -> CompletionResponse:
+        """Async completion endpoint for LLM."""
         pass
 
     @abstractmethod
     async def astream_chat(
         self, messages: Sequence[ChatMessage], **kwargs: Any
     ) -> ChatResponseGen:
+        """Async streaming chat endpoint for LLM."""
         pass
 
     @abstractmethod
     async def astream_complete(
         self, prompt: str, **kwargs: Any
     ) -> CompletionResponseGen:
+        """Async streaming completion endpoint for LLM."""
         pass