+++ llama_index/llms/openai.py
@@ -38,6 +38,7 @@ class OpenAI(LLM, BaseModel):
         return LLMMetadata(
             context_window=openai_modelname_to_contextsize(self.model),
             num_output=self.max_tokens or -1,
+            is_chat_model=self._is_chat_model,
         )
 
     def chat(self, messages: Sequence[ChatMessage], **kwargs: Any) -> ChatResponse: