+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+### Bug Fixes/ Nits
+- Improved chat refine template (#6645)
+
 ### Breaking/Deprecated API Changes
 - Change `BaseOpenAIAgent` to use `llama_index.llms.OpenAI`. Adjust `chat_history` to use `List[ChatMessage]]` as type.
 - Remove (previously deprecated) `llama_index.langchain_helpers.chain_wrapper` module.